api_yamdb/requirements.txt
psycopg2-binary==2.8.6
asgiref==3.2.10
atomicwrites==1.4.1
attrs==22.2.0
colorama==0.4.6
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
py==1.11.0
PyJWT==2.1.0
pytest==6.2.5
pytest-django==4.5.2
pytest-pythonpath==0.7.4
pytz==2020.1
sqlparse==0.3.1
toml==0.10.2
